Add files via upload
VIRTUAL PRIVATE CLOUD 
created a VPC in AWS that includes a public subnet for external access and a private subnet for internal resources, both hosting EC2 instances. A NAT gateway is set up to allow the private subnet to access the internet securely. Security groups are configured to manage inbound and outbound traffic for these instances.